+++ src/main/java/kitchenpos/domain/tablegroup/TableGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.table;
+package kitchenpos.domain.tablegroup;
 
 import kitchenpos.domain.order.OrderTable;
